Centralised the capybara screenshot code - was failing in cucumber
