+++ features/support/capybara_screenshot.rb
@@ -1,5 +1,2 @@
 require 'capybara-screenshot/cucumber'
-Capybara::Screenshot.prune_strategy = :keep_last_run
-Capybara::Screenshot.register_driver(:chrome) do |driver, path|
-  driver.browser.save_screenshot(path)
-end
+require_relative '../../test_common/capybara/capybara_screenshot'
